meet_schedule/requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==3.1.1
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
pip==21.3.1
promise==2.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==1.7.1
pytz==2021.3
Rx==1.6.1
setuptools==58.3.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
wheel==0.37.1

